Simplify comparison

+++ src/Arborescence.Search/EnumerableAStar/EnumerableAStar_6.cs
@@ -36,7 +36,7 @@ namespace Arborescence.Search
             if (vertexCount < 0)
                 throw new ArgumentOutOfRangeException(nameof(vertexCount));
 
-            if (unchecked((uint)source >= (uint)vertexCount))
+            if (unchecked((uint)source >= vertexCount))
                 yield break;
 
             byte[] colorMap = ArrayPool<byte>.Shared.Rent(vertexCount);
